% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{tsg_annotate}
\alias{tsg_annotate}
\title{Run TheSpliceGirls}
\usage{
tsg_annotate(splices)
}
\arguments{
\item{splices}{A vector of leafcutter transcript splicing IDs, i.e. "11:2384682:2390412:clu_12345_+"}
}
\value{
A dataframe containing the annotations
}
\description{
This function runs TheSpliceGirls annotation.
}
