% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{tsg_update_cache}
\alias{tsg_update_cache}
\title{Update TheSpliceGirls cache}
\usage{
tsg_update_cache(gencode = "46")
}
\arguments{
\item{gencode}{Desired gencode version (default: 46)}
}
\description{
This function updates the cache for TheSpliceGirls.
}
